メソッドをまとめるストラクトを追加
余計なデバッグを削除
